# docs/swagger.yaml
basePath: /api
definitions:
  handlers.Subscription:
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  handlers.SubscriptionCreateRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        description: '"07-2025"'
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для управления подписками
  title: SubServices API
  version: "1.0"
paths:
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Фильтрация по периоду
      parameters:
      - description: Начало периода (MM-YYYY)
        in: query
        name: from
        type: string
      - description: Конец периода (MM-YYYY)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.Subscription'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Получить подписки за период
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создаёт новую подписку. start_date и end_date передаются в формате
        MM-YYYY
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/handlers.SubscriptionCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Удаляет подписку по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            type: string
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      description: Возвращает подписку по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Subscription'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Получить подписку
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Полностью обновляет подписку по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: string
      - description: Новые данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/handlers.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Subscription'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Обновить подписку
      tags:
      - subscriptions
swagger: "2.0"